+++ integration_tests/test_ibc_rly.py
@@ -93,7 +93,7 @@ def coin_spent(spender, amt, denom):
 def distribute_fee(receiver, fee):
     return {
         "receiver": receiver,
-        "fee": keccak(text=fee),
+        "fee": fee,
     }
 
 
@@ -101,7 +101,7 @@ def fungible(dst, src, amt, denom):
     return {
         "receiver": dst,
         "sender": src,
-        "denom": keccak(text=denom),
+        "denom": denom,
         "amount": amt,
     }
 
@@ -123,9 +123,11 @@ def burn(burner, amt, denom):
 
 def recv_packet(seq, src, dst, amt, denom):
     return {
-        "packetSequence": keccak(text=f"{seq}"),
+        "packetSequence": seq,
         "packetSrcPort": keccak(text="transfer"),
         "packetSrcChannel": keccak(text=channel),
+        "packetSrcPortInfo": "transfer",
+        "packetSrcChannelInfo": channel,
         "packetDstPort": "transfer",
         "packetDstChannel": channel,
         "connectionId": "connection-0",
@@ -141,9 +143,11 @@ def recv_packet(seq, src, dst, amt, denom):
 
 def acknowledge_packet(seq):
     return {
-        "packetSequence": keccak(text=f"{seq}"),
+        "packetSequence": seq,
         "packetSrcPort": keccak(text="transfer"),
         "packetSrcChannel": keccak(text=channel),
+        "packetSrcPortInfo": "transfer",
+        "packetSrcChannelInfo": channel,
         "packetDstPort": "transfer",
         "packetDstChannel": channel,
         "connectionId": "connection-0",
@@ -152,15 +156,17 @@ def acknowledge_packet(seq):
 
 def denom_trace(denom):
     return {
-        "denom": keccak(text=denom),
+        "denom": denom,
     }
 
 
 def write_ack(seq, src, dst, amt, denom):
     return {
-        "packetSequence": keccak(text=f"{seq}"),
+        "packetSequence": seq,
         "packetSrcPort": keccak(text="transfer"),
         "packetSrcChannel": keccak(text=channel),
+        "packetSrcPortInfo": "transfer",
+        "packetSrcChannelInfo": channel,
         "packetDstPort": "transfer",
         "packetDstChannel": channel,
         "connectionId": "connection-0",
@@ -209,7 +215,7 @@ def get_send_packet_seq(
         events = parse_events_rpc(res["events"])
         target = events.get("send_packet")
         if target and target["packet_sequence"]:
-            return target["packet_sequence"]
+            return int(target["packet_sequence"])
     return None
 
 
@@ -220,7 +226,7 @@ def filter_logs_since(w3, start, name, seq):
         {
             "fromBlock": start,
             "address": [CONTRACT],
-            "topics": [topic, "0x" + keccak(text=f"{seq}").hex()],
+            "topics": [topic, "0x{:064x}".format(seq)],
         }
     )