+++ x/cronos/events/types/types.go
@@ -1,6 +1,8 @@
 package types
 
 const (
-	EventTypeSubmitMsgsResult = "submit_msgs_result"
-	AttributeKeySeq           = "seq"
+	EventTypeSubmitMsgsResult  = "submit_msgs_result"
+	AttributeKeySeq            = "seq"
+	AttributeKeySrcPortInfo    = "packet_src_port_info"
+	AttributeKeySrcChannelInfo = "packet_src_channel_info"
 )